Improve and clean tst_keras_tensorflow.py
This test now computes the memory usage of the tested models.
It also makes sure that the other backends are not visible in sys.modules
(if they are imported, an exception will be thrown).
The list of tested models is directly given by the Classifiers class, which
ensures that all the models that the factory can build are tested.