Create tests for keras with torch and jax backends
They are copies of tst_keras_tensorflow BUT some models cannot be created
with these configurations: efficientnet* models use preprocessing layers
like Rescale, and these layers necessarily use tensorflow. This is a known
limitation to the Keras team and as far as I know, there is no plan to
mitigate this issue for other backends.