Remove tst_tfkeras and tfkeras
The purpose of these files is to make sure the library behaves well when
using tensorflow.keras.
Starting with tensorflow 2.16.1, here is what we can read on the
tensorflow.keras README.md:

STOP!
This folder contains the legacy Keras code which is stale and about to be
deleted. The current Keras code lives in github/keras-team/keras.
Please do not use the code from this folder.

The message is pretty clear: do not use tensorflow.keras anymore.